refactor: consolidate how we do DNS lookups
