+++ pkg/x/dsl/api.go
@@ -14,9 +14,9 @@ func DNSLookupParallel(functions ...any) []any {
 	return EncodeFunctionList(&dnsLookupParallelTemplate{}, functions)
 }
 
-// Getaddrinfo resolves a domain name using getaddrinfo.
-func Getaddrinfo() []any {
-	return []any{(&getaddrinfoTemplate{}).Name()}
+// DNSLookupGetaddrinfo resolves a domain name using getaddrinfo.
+func DNSLookupGetaddrinfo() []any {
+	return []any{(&dnsLookupGetaddrinfoTemplate{}).Name()}
 }
 
 // MakeEndpointList transforms IP addresses to a list of endpoints.
@@ -54,9 +54,9 @@ func String(value string) []any {
 	return EncodeFunctionScalar(&stringTemplate{}, value)
 }
 
-// UDPResolver constructs an UDP resolver using the given endpoint address. For IPv4
+// DNSLookupUDP constructs an UDP resolver using the given endpoint address. For IPv4
 // endpoints use the "<address>:<port>" pattern (e.g., "8.8.8.8.8:53"). Make sure you
 // quote the address (e.g., "[2001:4860:4860::8844]:53") for IPv6 endpoints.
-func UDPResolver(value string) []any {
-	return EncodeFunctionScalar(&udpResolverTemplate{}, value)
+func DNSLookupUDP(value string) []any {
+	return EncodeFunctionScalar(&dnsLookupUDPTemplate{}, value)
 }
